requirements.txt
Django==1.6.5
South==0.8.4
django-bootstrap3==4.4.0
django-debug-toolbar==1.2.1
django-discover-runner==1.0
django-forkit==0.9.5
django-model-utils==2.0.3
django-nested-formset==0.1.2
django-picklefield==0.3.1
django-vanilla-views==1.0.3
gnureadline==6.3.3
ipython==2.0.0
readline==6.2.4.1
rebar==0.1
selenium==2.41.0
six==1.6.1
sqlparse==0.1.11
sympy==0.7.5
